refactor: rename classes and namespaces
